backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20     # file uploads
websockets==13.1              # WebSocket support

# ── Configuration & Validation ─────────────────────────────────────────────────
pydantic-settings==2.7.0
email-validator==2.2.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# ── Vector Store ───────────────────────────────────────────────────────────────
chromadb==0.6.3

# ── HTTP Client ────────────────────────────────────────────────────────────────
httpx==0.28.1

# ── Auth & Security ────────────────────────────────────────────────────────────
PyJWT==2.10.1
bcrypt==4.3.0
ldap3==2.9.1                  # Pure-Python LDAP client for Active Directory

# ── Embeddings (local, offline) ────────────────────────────────────────────────
sentence-transformers==3.3.1  # all-MiniLM-L6-v2 local embedding model

# ── Document Parsing ───────────────────────────────────────────────────────────
pypdf==4.3.1                  # PDF text extraction
python-docx==1.1.2            # DOCX text extraction
openpyxl==3.1.5               # XLSX text extraction
aiofiles==24.1.0              # async file I/O

# ── File Watching ───────────────────────────────────────────────────────────────
watchdog==6.0.0               # Monitor directories for new/modified files

# ── System Metrics ──────────────────────────────────────────────────────────────
psutil==6.1.1                 # CPU/RAM/disk metrics for /health/services endpoint

# ── Knowledge Source Connectors ─────────────────────────────────────────────────
smbprotocol==1.15.0           # SMB/CIFS file server connector
exchangelib==5.6.0            # Exchange/Outlook EWS connector
msal==1.32.3                  # Microsoft SharePoint / Graph API auth (MSAL)
boto3==1.37.23                # S3-compatible object storage connector
